turn off DEBUG-UART && pin && laser && display

+++ Project/CODE/motor.h
@@ -12,9 +12,9 @@
 
 typedef struct motor_param_t
 {
-    int64_t total_encoder;
-    int64_t target_encoder;
-    int16_t encoder_raw;
+    float total_encoder;
+    float target_encoder;
+    float encoder_raw;
     float encoder_speed; //Measured speed
     float target_speed;
     int32_t duty;         //Motor PWM duty
